refactor: 자료형 변경
* Integer 자료형의 범위를 넘어선 계산을 하기 위해 BigInteger 클래스를 사용하였다.

+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -3,6 +3,7 @@ package lotto.domain;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -83,6 +84,6 @@ public class LottoResultTest {
     @Test
     void 총_수익률_계산() {
         LottoResult lottoResult = new LottoResult(winningLotto, lottos);
-        assertThat(40032100).isEqualTo(lottoResult.getEarningsRate());
+        assertThat(BigInteger.valueOf(40032100)).isEqualTo(lottoResult.getEarningsRate());
     }
 }
